+++ desc/particles.py
@@ -14,16 +14,10 @@ from diffrax import (
     Tsit5,
     diffeqsolve,
 )
-from scipy.constants import (
-    Boltzmann,
-    electron_mass,
-    elementary_charge,
-    epsilon_0,
-    hbar,
-    proton_mass,
-)
+from scipy.constants import Boltzmann, elementary_charge, proton_mass
 
-from desc.backend import jit, jnp, tree_map, vmap
+from desc.backend import jax, jit, jnp, tree_map
+from desc.batching import vmap_chunked
 from desc.compute.utils import _compute as compute_fun
 from desc.compute.utils import get_profiles, get_transforms
 from desc.derivatives import Derivative
@@ -132,6 +126,12 @@ class VacuumGuidingCenterTrajectory(AbstractTrajectoryModel):
         integration. Thus, it is not implemented. For that case, we recommend converting
         the final output to "lab" frame after the integration is done using
         Equilibrium.map_coordinates method.
+
+        If anytime during the integration, the particle's rho coordinate becomes
+        smaller than 1e-6, the magnetic field is computed at rho = 1e-6 to avoid
+        numerical issues with the rho = 0. Note that particle position doesn't have
+        discontinuity due to this, only the magnetic field is computed at a different
+        point.
     """
 
     vcoords = ["vpar"]
@@ -203,13 +203,21 @@ class VacuumGuidingCenterTrajectory(AbstractTrajectoryModel):
         iota profile, it must be passed as a keyword argument in kwargs. In that case,
         params should also contain i_l, which is the iota profile parameters.
         """
-        rho, theta, zeta, vpar = x
+        xp, yp, zeta, vpar = x
+        # we use cartesian-like coordinates xp and yp which are x=rho*cos(theta) and
+        # y=rho*sin(theta) for integration, but convert them to flux coordinates for
+        # compute functions. This way, we don't have the problem of terminating the
+        # integration when rho<0, but actually the particle is still in the plasma
+        # volume.
+        rho = jnp.sqrt(xp**2 + yp**2)
+        theta = jnp.arctan2(yp, xp)
+        # compute functions are not correct for very small rho
+        rho = jnp.where(rho < 1e-6, 1e-6, rho)
         iota = kwargs.get("iota", None)
         grid = Grid(
             jnp.array([rho, theta, zeta]).T,
             spacing=jnp.zeros((3,)).T,
             jitable=True,
-            sort=False,
         )
         data_keys = [
             "B",
@@ -237,8 +245,12 @@ class VacuumGuidingCenterTrajectory(AbstractTrajectoryModel):
         rhodot = dot(Rdot, data["e^rho"])
         thetadot = dot(Rdot, data["e^theta"])
         zetadot = dot(Rdot, data["e^zeta"])
+        # get the derivative for cartesian-like coordinates
+        xpdot = rhodot * jnp.cos(theta) - rho * thetadot * jnp.sin(theta)
+        ypdot = rhodot * jnp.sin(theta) + rho * thetadot * jnp.cos(theta)
+        # derivative the parallel velocity
         vpardot = -mu / m * dot(data["b"], data["grad(|B|)"])
-        dxdt = jnp.array([rhodot, thetadot, zetadot, vpardot]).reshape(x.shape)
+        dxdt = jnp.array([xpdot, ypdot, zetadot, vpardot]).reshape(x.shape)
         return dxdt.squeeze()
 
     def _compute_lab_coordinates(self, x, field, params, m, q, mu, **kwargs):
@@ -274,174 +286,6 @@ class VacuumGuidingCenterTrajectory(AbstractTrajectoryModel):
 
         vpardot = jnp.atleast_1d(-mu / m * dot(b, grad_B))
         dxdt = jnp.concatenate([Rdot, vpardot]).reshape(x.shape)
-        return dxdt.squeeze()
-
-
-class SlowingDownGuidingCenterTrajectory(AbstractTrajectoryModel):
-    """Guiding center trajectories with slowing down on electrons and main ions.
-
-    Solves the following ODEs,
-
-    d𝐑/dt = v∥ 𝐛 + (m / q B²) ⋅ (v∥² + 1/2 v⊥²) ( 𝐛 × ∇B )
-
-    dv∥/dt = − ((v² - v∥²) / 2B) ( 𝐛 ⋅ ∇B )
-
-    dv/dt = - v / τₛ (1 + v_c³ / v³)
-
-    where 𝐁 is the magnetic field vector at position 𝐑, B is the magnitude of
-    the magnetic field, 𝐛 is the unit magnetic field 𝐁/B, τₛ is the Spitzer
-    ion–electron momentum exchange time and v_c is the critical velocity associated
-    with the critical energy at which the velocity reduction transitions from
-    nearly exponential (drag on electrons) to significantly steeper (drag on ions).
-    τₛ and v_c are defined as follows:
-
-    τₛ = (mᵢ (4πϵ₀)² 3mₑ¹ᐟ² Tₑ³ᐟ²) / (mₑ 4√(2π) nₑ Zᵢ² e⁴ lnΛ)
-
-    v_c = [ (3√π / 4) (mₑ / mᵢ) ]¹ᐟ³ v_Tₑ
-
-    v_Tₑ = √(2 Tₑ / mₑ)
-
-    See ref [1]_ Eq. (1-4) for definitions, and other references [2]_ [3]_ [4]_
-    for the details.
-
-    Works only in flux coordinates corresponding to {rho, theta, zeta}. Particle
-    tracing can be performed with an Equilibrium object, which must have electron
-    temperature Te and electron density ne defined.
-
-    Parameters
-    ----------
-    m_eff : float
-        Effective mass of the plasma main ions, in units of proton mass. Default is 2.5,
-        for a 50/50 DT plasma
-    Z_eff : float
-        Effective charge of the plasma main ions, in units of elementary charge.
-        Default is 1, for H/D/T plasmas.
-
-    References
-    ----------
-    .. [1] McMillan M, Lazerson S A. "BEAMS3D neutral beam injection model"
-       Plasma Physics and Controlled Fusion (2014).
-    .. [2] Callen J D, "Fundamentals of Plasma Physics (Lecture Notes)" (Madison, WI:
-       University of Wisconsin Press) (2003)
-    .. [3] Fowler R H, Morris R N, Rome J A and Hanatani K, "Neutral beam injection
-       benchmark studies for stellarators/heliotrons", Nucl. Fusion 30 997–1010 (1990)
-    .. [4] Rosenbluth M N, MacDonald W M and Judd D L, "Fokker–Planck equation for an
-       inverse-square force", Phys.Rev. 107 1–6 (1957)
-
-    """
-
-    _Z_eff: float
-    _m_eff: float
-    vcoords = ["vpar", "v"]
-    args = ["m", "q"]
-
-    def __init__(self, frame="flux", m_eff=2.5, Z_eff=1.0):
-        assert frame == "flux"
-        self._frame = frame
-        self._Z_eff = Z_eff
-        self._m_eff = m_eff
-
-    @property
-    def Z_eff(self):
-        """Coordinate frame of the model."""
-        return self._Z_eff
-
-    @property
-    def m_eff(self):
-        """Coordinate frame of the model."""
-        return self._m_eff
-
-    @property
-    def frame(self):
-        """Coordinate frame of the model."""
-        return self._frame
-
-    @jit
-    def vf(self, t, x, args):
-        """RHS of guiding center trajectories without collisions or slowing down.
-
-        ``vf`` method corresponds to the ``vf`` method in diffrax.AbstractTerm class
-        and must have the same name and signature.
-
-        Parameters
-        ----------
-        t : float
-            Time to evaluate RHS at.
-        x : jax.Array, shape(N,5)
-            Position of particle in phase space (rho, theta, zeta, vpar, v).
-        args : tuple
-            Should include the arguments needed by the model, (m, q) as
-            an array, Equilibrium object, params and any additional keyword
-            arguments needed for magnetic field computation, i.e. iota profile for
-            the Equilibrium. Note: if Equilibrium does not have iota profile,
-            params dictionary must contain i_l, which is the iota profile parameters.
-
-        Returns
-        -------
-        dx : jax.Array, shape(N,5)
-            Velocity of particles in phase space.
-        """
-        rho, theta, zeta, vpar, v = x
-        model_args, eq, params, kwargs = args
-        m, q = model_args
-        iota = kwargs.get("iota", None)
-
-        assert (
-            eq.Te_l.size > 0
-        ), "Equilibrium must have electron temperature Te defined."
-        assert eq.ne_l.size > 0, "Equilibrium must have electron density ne defined."
-
-        grid = Grid(
-            jnp.array([rho, theta, zeta]).T,
-            spacing=jnp.zeros((3,)).T,
-            jitable=True,
-            sort=False,
-        )
-        data_keys = [
-            "B",
-            "|B|",
-            "grad(|B|)",
-            "e^rho",
-            "e^theta",
-            "e^zeta",
-            "b",
-            "Te",
-            "ne",
-        ]
-
-        transforms = get_transforms(data_keys, eq, grid, jitable=True)
-        profiles = get_profiles(data_keys, eq, grid)
-        if iota is not None:
-            profiles["iota"] = iota
-        data = compute_fun(
-            eq,
-            data_keys,
-            params,
-            transforms,
-            profiles,
-        )
-
-        # slowing eqns from McMillan, Matthew, and Samuel A. Lazerson. "BEAMS3D
-        # neutral beam injection model." Plasma Physics and Controlled Fusion (2014)
-        tau_s = slowing_down_time(data["Te"], data["ne"], self.m_eff, self.Z_eff)
-        vc = slowing_down_critical_velocity(data["Te"], self.m_eff)
-
-        # derivative of the guiding center position in R, phi, Z coordinates
-        # TODO: Is this correct? Check
-        Rdot = vpar * data["b"] + (
-            (m / q / data["|B|"] ** 2)
-            * (vpar**2 + 0.5 * (v**2 - vpar**2))
-            * cross(data["b"], data["grad(|B|)"])
-        )
-        # take dot product for rho, theta and zeta coordinates
-        rhodot = dot(Rdot, data["e^rho"])
-        thetadot = dot(Rdot, data["e^theta"])
-        zetadot = dot(Rdot, data["e^zeta"])
-        vpardot = (
-            -(v**2 - vpar**2) / (2 * data["|B|"]) * dot(data["b"], data["grad(|B|)"])
-        )
-        vdot = -v / tau_s * (1 + vc**3 / v**3)
-        dxdt = jnp.array([rhodot, thetadot, zetadot, vpardot, vdot]).reshape(x.shape)
         return dxdt.squeeze()
 
 
@@ -531,9 +375,17 @@ def _compute_modB(x, field, params, **kwargs):
             NFP=field.NFP,
         )
         profiles = get_profiles("|B|", field, grid)
+        transforms = get_transforms("|B|", field, grid, jitable=True)
         if "iota" in kwargs:
             profiles["iota"] = kwargs["iota"]
-        return field.compute("|B|", params=params, grid=grid, profiles=profiles)["|B|"]
+        return compute_fun(
+            field,
+            "|B|",
+            params=params,
+            grid=grid,
+            profiles=profiles,
+            transforms=transforms,
+        )["|B|"]
     source_grid = kwargs.pop("source_grid", None)
     return jnp.linalg.norm(
         field.compute_magnetic_field(x, params=params, source_grid=source_grid), axis=-1
@@ -846,7 +698,13 @@ class CurveParticleInitializer(AbstractParticleInitializer):
         data = self.curve.compute(["x_s", "x", "ds"], grid=self.grid)
         # length of the line segment at each grid point
         sqrtg = jnp.linalg.norm(data["x_s"], axis=-1) * data["ds"]
-        self._chosen_idxs = _find_random_indices(sqrtg, self.N, seed=self.seed)
+        self._chosen_idxs = jax.random.choice(
+            key=jax.random.PRNGKey(self.seed),
+            a=sqrtg.shape[0],
+            shape=(self.N,),
+            replace=True,
+            p=sqrtg / sqrtg.sum(),
+        )
 
         # positions of the selected nodes in R, phi, Z coordinates
         x = data["x"][self._chosen_idxs, :]
@@ -972,8 +830,13 @@ class SurfaceParticleInitializer(AbstractParticleInitializer):
         # surface area for each grid point
         # include spacing (dt*dz) to account for non-uniform grids
         sqrtg = data["|e_theta x e_zeta|"] * self.grid.spacing[:, 1:].prod(axis=-1)
-        self._chosen_idxs = _find_random_indices(sqrtg, self.N, seed=self.seed)
-
+        self._chosen_idxs = jax.random.choice(
+            key=jax.random.PRNGKey(self.seed),
+            a=sqrtg.shape[0],
+            shape=(self.N,),
+            replace=True,
+            p=sqrtg / sqrtg.sum(),
+        )
         # eq and surface might not have the same theta definition, so we will do a
         # root finding to find the correct theta and zeta coordinates from R, phi, Z
         # coordinates. We will use the surface's rho, theta, zeta coordinates as an
@@ -1029,35 +892,10 @@ class SurfaceParticleInitializer(AbstractParticleInitializer):
         )
 
 
-def _find_random_indices(sqrtg, N, seed):
-    """Find random indices for sampling particles on a surface or curve."""
-    # probability of particle generation in a given grid point is proportional to
-    # its volume/area/length, which is sqrtg. Normalize sqrtg for random number
-    # generation limit of 1
-    sqrtg /= sqrtg.max()
-    nattempts = 10 * N
-    rng = np.random.default_rng(seed=seed)
-    accept = []
-    # loop until choosing exactly N indices
-    # note: generated random integers might repeat, but we will still keep them
-    # So, the final list of indices might have duplicates, meaning that a grid point
-    # spawned multiple particles.
-    while len(accept) < N:
-        idxs = rng.integers(0, sqrtg.shape[0], size=(nattempts,))
-        # note: probability of selecting a number <0.3 in range [0, 1] is 30%
-        accept = np.where(rng.uniform(0, 1, size=(idxs.shape[0],)) < sqrtg[idxs])[0]
-        # choose random N of the accepted indices
-        accept = accept[rng.integers(0, accept.shape[0], size=(N,))]
-        idxs = idxs[accept]
-
-    return np.sort(idxs)
-
-
 def trace_particles(
     field,
-    y0,
+    initializer,
     model,
-    model_args,
     ts,
     params=None,
     stepsize_controller=None,
@@ -1070,24 +908,19 @@ def trace_particles(
     adjoint=RecursiveCheckpointAdjoint(),
     bounds=None,
     event=None,
+    chunk_size=None,
     options={},
 ):
     """Trace charged particles in an equilibrium or external magnetic field.
 
-    If this function will be used in an optimization loop, it is recommended to
-    pass a value for all the keyword arguments to prevent recompilation.
+    For jit friendly version of this function, see `_trace_particles`.
 
     Parameters
     ----------
     field : MagneticField or Equilibrium
         Source of magnetic field to integrate
-    y0 : array-like
-        Initial particle positions and velocities, stacked in horizontally [x0, v0].
-        The first output of ``ParticleInitializer.init_particles``.
-    model_args : array-like
-        Additional arguments needed by the model, such as mass, charge, and
-        magnetic moment (mv⊥²/2|B|) of each particle. The second output of
-        ``ParticleInitializer.init_particles``.
+    initializer : AbstractParticleInitializer
+        Particle initializer
     ts : array-like
         Strictly increasing array of time values where output is desired.
     model : AbstractTrajectoryModel
@@ -1129,6 +962,9 @@ def trace_particles(
         event function will be used, which stops integration when particles leave the
         bounds. If integration is stopped by the event, the output will contain NaN
         values for the points outside the bounds.
+    chunk_size : int, optional
+        Chunk size for integration over particles. If not provided, the integration will
+        be done over all particles at once without chunking.
     options : dict, optional
         Additional keyword arguments to pass to the field computation, such as
             - iota : Profile
@@ -1146,6 +982,71 @@ def trace_particles(
         will depend on ``model.vcoords``.
 
     """
+    y0, model_args = initializer.init_particles(model, field)
+    return _trace_particles(
+        field=field,
+        y0=y0,
+        model=model,
+        model_args=model_args,
+        ts=ts,
+        params=params,
+        stepsize_controller=stepsize_controller,
+        saveat=saveat,
+        rtol=rtol,
+        atol=atol,
+        max_steps=max_steps,
+        min_step_size=min_step_size,
+        solver=solver,
+        adjoint=adjoint,
+        bounds=bounds,
+        event=event,
+        chunk_size=chunk_size,
+        options=options,
+    )
+
+
+def _trace_particles(
+    field,
+    y0,
+    model,
+    model_args,
+    ts,
+    params=None,
+    stepsize_controller=None,
+    saveat=None,
+    rtol=1e-8,
+    atol=1e-8,
+    max_steps=None,
+    min_step_size=1e-10,
+    solver=Tsit5(),
+    adjoint=RecursiveCheckpointAdjoint(),
+    bounds=None,
+    event=None,
+    chunk_size=None,
+    options={},
+):
+    """Trace charged particles in an equilibrium or external magnetic field.
+
+    This is the jit friendly version of the `trace_particles` function. For full
+    documentation, see `trace_particles`. This function takes the outputs of
+    `initializer.init_particles` as inputs, rather than the particle initializer
+    itself. There won't be any checks on the y0 and model_args inputs, so make sure
+    they are in the correct format. One can use this function in an objective, where
+    the initial positions and velocities of particles are computed in the `build`
+    method. If the objective requires initialization of particles at each iteration,
+    make sure that the initializer can work under jit compilation which is not the
+    case for all of them.
+
+    Parameters
+    ----------
+    y0 : array-like
+        Initial particle positions and velocities, stacked in horizontally [x0, v0].
+        The first output of ``initializer.init_particles``.
+    model_args : array-like
+        Additional arguments needed by the model, such as mass, charge, and
+        magnetic moment (mv⊥²/2|B|) of each particle. The second output of
+        ``initializer.init_particles``.
+    """
     if not params:
         params = field.params_dict
 
@@ -1161,25 +1062,39 @@ def trace_particles(
         if isinstance(field, Equilibrium):
             bounds = bounds.at[0, 1].set(1.0)  # rho bounds for flux coordinates
 
-    def default_terminating_event(t, y, args, **kwargs):
-        i_out = jnp.logical_or(y[0] < bounds[0, 0], y[0] > bounds[0, 1])
-        j_out = jnp.logical_or(y[1] < bounds[1, 0], y[1] > bounds[1, 1])
+    def default_event(t, y, args, **kwargs):
+        if isinstance(field, Equilibrium):
+            i = jnp.sqrt(y[0] ** 2 + y[1] ** 2)
+            j = jnp.arctan2(y[1], y[0])
+        else:
+            i = y[0]
+            j = y[1]
+        i_out = jnp.logical_or(i < bounds[0, 0], i > bounds[0, 1])
+        j_out = jnp.logical_or(j < bounds[1, 0], j > bounds[1, 1])
         k_out = jnp.logical_or(y[2] < bounds[2, 0], y[2] > bounds[2, 1])
         return jnp.logical_or(i_out, jnp.logical_or(j_out, k_out))
 
-    event = Event(default_terminating_event) if event is None else event
+    event = Event(default_event) if event is None else event
     max_steps = (
         max_steps
         if max_steps is not None
         else int((ts[1] - ts[0]) / min_step_size * 100)
     )
+
+    # convert cartesian-like for integration in flux coordinates
+    if isinstance(field, Equilibrium):
+        xp = y0[:, 0] * jnp.cos(y0[:, 1])
+        yp = y0[:, 0] * jnp.sin(y0[:, 1])
+        y0 = y0.at[:, 0].set(xp)
+        y0 = y0.at[:, 1].set(yp)
+
     with warnings.catch_warnings():
         warnings.filterwarnings("ignore", message="unhashable type")
-        # we only want to map over initial positions and particle arguments,
-        # other arguments are there to prevent closing over them, and hence
-        # reduce the number of recompilations in, for example, an optimization
-        # loop. Note: vmap with keyword arguments is weird, not using it for now
-        yt = vmap(_intfun_wrapper, in_axes=(0, 0) + 12 * (None,))(
+        # we only want to map over initial positions and particle arguments
+        # Note: vmap with keyword arguments is weird, not using it for now
+        yt = vmap_chunked(
+            _intfun_wrapper, in_axes=(0, 0) + 12 * (None,), chunk_size=chunk_size
+        )(
             y0,
             model_args,
             field,
@@ -1200,6 +1115,13 @@ def trace_particles(
 
     x = yt[:, :, :3]
     v = yt[:, :, 3:]
+
+    # convert back to flux coordinates
+    if isinstance(field, Equilibrium):
+        rho = jnp.sqrt(x[:, :, 0] ** 2 + x[:, :, 1] ** 2)
+        theta = jnp.arctan2(x[:, :, 1], x[:, :, 0])
+        x = x.at[:, :, 0].set(rho)
+        x = x.at[:, :, 1].set(theta)
 
     return x, v
 
@@ -1223,7 +1145,7 @@ def _intfun_wrapper(
     """Wrapper for the integration function for vectorized inputs.
 
     Defining a lambda function inside the ``trace_particles`` function leads
-    to recompilations.
+    to recompilations, so instead we define the wrapper here.
     """
     return diffeqsolve(
         terms=model,
@@ -1239,190 +1161,3 @@ def _intfun_wrapper(
         adjoint=adjoint,
         event=event,
     ).ys
-
-
-def gc_radius(vperp, modB, m, q):
-    """Radius of guiding center orbit.
-
-    Parameters
-    ----------
-    vperp : array-like
-        Magnitude of perpendicular velocity, in m/s
-    modB : array-like
-        Magnitude of magnetic field, in T.
-    m : array-like
-        Mass of particle, in kg.
-    q : array-like
-        Charge of particle, in C.
-
-    Returns
-    -------
-    rho : jax.Array
-        Gyroradius, in meters.
-    """
-    return m * vperp / (jnp.abs(q) * modB)
-
-
-def slowing_down_time(Te, ne, m_eff=2.5, Z_eff=1):
-    """Slowing down time for fast ion friction with electrons.
-
-    Parameters
-    ----------
-    Te : array-like
-        Electron temperature, in eV
-    ne : array-like
-        Electron density, in particles/m^3
-    m_eff : array-like
-        Effective mass of the plasma main ions, in units of proton mass. Default is 2.5,
-        for a 50/50 DT plasma
-    Z_eff : array-like
-        Effective charge of the plasma main ions, in units of elementary charge.
-        Default is 1, for H/D/T plasmas.
-
-    Returns
-    -------
-    tau_s : jax.Array
-        Slowing down time, in seconds.
-    """
-    me = electron_mass
-    mi = m_eff * proton_mass
-
-    lnlambda = coulomb_logarithm(ne, ne / Z_eff, Te, Te, m_eff, Z_eff)
-
-    tau_s = (
-        (mi / me)
-        * (4 * jnp.pi * epsilon_0) ** 2
-        / (4 * jnp.sqrt(2 * jnp.pi))
-        * (3 * me ** (1 / 2) * (Te * JOULE_PER_EV) ** (3 / 2))
-        / (ne * Z_eff**2 * elementary_charge**4 * lnlambda)
-    )
-    return tau_s
-
-
-def slowing_down_critical_velocity(Te, m_eff=2.5):
-    """Critical velocity for transition from slowing down on electrons to ions.
-
-    For fast ion speeds above the critical velocity the particles primarily slow down
-    due to friction with electrons, which reduces the particle speed exponentially.
-    Below the critical velocity friction with the ions dominates and the slowing down
-    is super-exponential.
-
-    Parameters
-    ----------
-    Te : array-like
-        Electron temperature, in eV
-    m_eff : array-like
-        Effective mass of the plasma main ions, in units of proton mass. Default is 2.5,
-        for a 50/50 DT plasma
-
-    Returns
-    -------
-    vc : jax.Array
-        Critical velocity.
-    """
-    me = electron_mass
-    mi = m_eff * proton_mass
-    vTe = jnp.sqrt(2 * Te * JOULE_PER_EV / me)
-    vc = (3 * jnp.sqrt(jnp.pi) / 4 * me / mi) ** (1 / 3) * vTe
-    return vc
-
-
-def coulomb_logarithm(ne, ni, Te, Ti, m_eff, Z_eff) -> float:
-    """Coulomb logarithm for collisions between species a and b.
-
-    Parameters
-    ----------
-    ne, ni : float
-        Density of electrons and ions in particles/m^3.
-    Te, Ti : float
-        Temperature of electrons and ions in eV.
-    m_eff : float
-        Effective mass of ions, in units of proton mass.
-    Z_eff : float
-        Effective charge of ions, in units of elementary charge.
-
-    Returns
-    -------
-    log(lambda) : float
-
-    """
-    bmin, bmax = impact_parameter(ne, ni, Te, Ti, m_eff, Z_eff)
-    return jnp.log(bmax / bmin)
-
-
-def impact_parameter(ne, ni, Te, Ti, m_eff, Z_eff) -> float:
-    """Impact parameters for classical Coulomb collision.
-
-    Parameters
-    ----------
-    ne, ni : float
-        Density of electrons and ions in particles/m^3.
-    Te, Ti : float
-        Temperature of electrons and ions in eV.
-    m_eff : float
-        Effective mass of ions, in units of proton mass.
-    Z_eff : float
-        Effective charge of ions, in units of elementary charge.
-    """
-    vte = jnp.sqrt(2 * Te * JOULE_PER_EV / electron_mass)
-    vti = jnp.sqrt(2 * Ti * JOULE_PER_EV / (m_eff * proton_mass))
-    bmin = jnp.maximum(
-        impact_parameter_perp(m_eff, Z_eff, vte, vti),
-        debroglie_length(m_eff, vte, vti),
-    )
-    bmax = debye_length(ne, ni, Te, Ti, Z_eff)
-    return bmin, bmax
-
-
-def impact_parameter_perp(m_eff, Z_eff, vte, vti) -> float:
-    """Distance of the closest approach for a 90° Coulomb collision.
-
-    Parameters
-    ----------
-    m_eff : float
-        Effective mass of ions, in units of proton mass.
-    Z_eff : float
-        Effective charge of ions, in units of elementary charge.
-    vte, vti : float
-        Thermal speeds of electrons and ions in m/s.
-    """
-    me = electron_mass
-    mi = m_eff * proton_mass
-    m_reduced = (mi * me) / (me + mi)
-    qe = -elementary_charge
-    qi = Z_eff * elementary_charge
-    return qe * qi / (4 * jnp.pi * epsilon_0 * m_reduced * (vte**2 + vti**2))
-
-
-def debroglie_length(m_eff, vte, vti) -> float:
-    """Thermal DeBroglie wavelength.
-
-    Parameters
-    ----------
-    m_eff : float
-        Effective mass of ions, in units of proton mass.
-    vte, vti : float
-        Thermal speeds of electrons and ions in m/s.
-    """
-    me = electron_mass
-    mi = m_eff * proton_mass
-    m_reduced = (mi * me) / (me + mi)
-    v_th = jnp.sqrt(vte**2 + vti**2)
-    return hbar / (2 * m_reduced * v_th)
-
-
-def debye_length(ne, ni, Te, Ti, Z_eff) -> float:
-    """Scale length for charge screening.
-
-    Parameters
-    ----------
-    ne, ni : float
-        Density of electrons and ions in particles/m^3
-    Te, Ti : float
-        Temperature of electrons and ions in eV
-    Z_eff : float
-        Effective charge of ions, in units of elementary charge.
-    """
-    den = ne * elementary_charge**2 / (Te * JOULE_PER_EV)
-    den += ni * Z_eff**2 * elementary_charge**2 / (Ti * JOULE_PER_EV)
-    return jnp.sqrt(epsilon_0 / den)